==> prior_experiments/1.issues_cleaning.ipynb <==
# Case Issues data cleaning

The labels we are most interested in are case issues and causes of action. This notebook outlines the steps undertook to generate and clean the data for case issues, to be used for model training and evaluation.

# Import libaries

import numpy as np
import pandas as pd

import re
from datetime import datetime

from sklearn.preprocessing import MultiLabelBinarizer

from clearinghouse_util import fetch_all_case_data

# Load the data

df = pd.read_json("0.clearinghouse_all_case_data.json")
df.head()

assert len(df) == df["case_id"].nunique()

# Initial EDA on the overall dataset

len(df)

df.columns

for each in ['case_status', 'case_state', 'court_name',
       'case_type', 'case_ongoing', 'case_special_collections',
       'case_causes_of_action', 'case_issues']:
    print("---")
    print(df[each].nunique())
    print(df[each].value_counts())

# Split up the case_issues to individual labels

def issue_split(text):
    text = str(text)
    parts = re.split(r',\s*(?![^()]*\))', text)
    return [part.strip() for part in parts]

df['issues'] = df['case_issues'].apply(issue_split)
df['issues_len'] = df['issues'].apply(len)
df['issues_len'].value_counts()

file_path = '1.clean_issues.json'
df.to_json(file_path)

mlb = MultiLabelBinarizer()
binary_matrix = mlb.fit_transform(df['issues'].to_list())
binary_matrix.shape

label_df = pd.DataFrame(binary_matrix, columns=mlb.classes_)
label_df.head()

label_df.sum().sort_values(ascending=False)

# Keep only the issues that are most frequently occurring (> 90 percentile)

label_counts = label_df.sum()

threshold = label_counts.quantile(0.90)
labels = label_counts[label_counts >= threshold]
labels.sort_values(ascending=False)

label_set = set(labels.index.tolist())
df['filtered_issues'] = df['issues'].apply(lambda labels: [label for label in labels if label in label_set])

df_filtered = df[df['filtered_issues'].str.len() > 0].reset_index(drop=True)
len(df_filtered)

file_path = '2.filtered_issues.json'
df_filtered.to_json(file_path)
