==> prior_experiments/3.train_test_split.ipynb <==
import numpy as np
import pandas as pd

from sklearn.model_selection import train_test_split

df = pd.read_json('3.docket_entry_issues.json')

metadata = pd.read_csv("clearinghouse.csv")

df = df.merge(metadata, how="left", on="case_id")
df.head()

len(df)

for each in ["case_status", "case_state", "court_name", "case_ongoing"]:
    print("---")
    print(each)
    print(df[each].nunique())
    print(df[each].value_counts())

train, test = train_test_split(df, test_size=0.3, random_state=42)

len(train)

for each in ["case_status", "case_state", "court_name", "case_ongoing"]:
    print("---")
    print(each)
    print(train[each].nunique())
    print(train[each].value_counts())

len(test)

for each in ["case_status", "case_state", "court_name", "case_ongoing"]:
    print("---")
    print(each)
    print(test[each].nunique())
    print(test[each].value_counts())

train = train[['case_id', 'filtered_issues', 'docket_entries', 'case_name',
       'case_status', 'case_state', 'court_name', 'case_ongoing']]
test = test[['case_id', 'filtered_issues', 'docket_entries', 'case_name',
       'case_status', 'case_state', 'court_name', 'case_ongoing']]

train.to_json("4a.train.json")
test.to_json("4b.test.json")
